pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "fastapi-boilerplate"
version = "0.1.0"
description = "FastAPI boilerplate (Clean Architecture) with Postgres, optional Redis, and healthchecks."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi[standard]>=0.115",
  "pydantic-settings>=2.0",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "python-ulid>=2.0",
  "structlog>=24.0",
  "redis>=5.0",
  "python-jose[cryptography]>=3.3",
  "passlib[bcrypt]>=1.7",
]

[dependency-groups]
dev = [
  "ruff>=0.5",
  "mypy>=1.10",
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
  "httpx>=0.27",
  "testcontainers[postgres,redis]>=4.0",
  "faker>=25.0",
  "asgi-lifespan>=2.1",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = [
  "B008",  # FastAPI: Depends(...) in signature defaults is idiomatic
  "E402",  # Our modules start with long docstrings/workflows before imports
  "UP017", # datetime.UTC alias is optional stylistic modernisation
  "UP042", # Enum patterns like `class Role(str, Enum)` are idiomatic
]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_optional = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = false
plugins = []
